# util: Pass limit to strutil.MergeSlices