--- a/pkg/strutil/merge.go
+++ b/pkg/strutil/merge.go
@@ -10,33 +10,39 @@ import (
 
 // MergeSlices merges a set of sorted string slices into a single ones
 // while removing all duplicates.
-func MergeSlices(a ...[]string) []string {
+// If limit is set, only the first limit results will be returned. 0 to disable.
+func MergeSlices(limit int, a ...[]string) []string {
 	if len(a) == 0 {
 		return nil
 	}
 	if len(a) == 1 {
-		return a[0]
+		return truncateToLimit(limit, a[0])
 	}
 	l := len(a) / 2
-	return mergeTwoStringSlices(MergeSlices(a[:l]...), MergeSlices(a[l:]...))
+	return mergeTwoStringSlices(limit, MergeSlices(limit, a[:l]...), MergeSlices(limit, a[l:]...))
 }
 
 // MergeUnsortedSlices behaves like StringSlices but input slices are validated
 // for sortedness and are sorted if they are not ordered yet.
-func MergeUnsortedSlices(a ...[]string) []string {
+// If limit is set, only the first limit results will be returned. 0 to disable.
+func MergeUnsortedSlices(limit int, a ...[]string) []string {
 	for _, s := range a {
 		if !sort.StringsAreSorted(s) {
 			sort.Strings(s)
 		}
 	}
-	return MergeSlices(a...)
+	return MergeSlices(limit, a...)
 }
 
-func mergeTwoStringSlices(a, b []string) []string {
+func mergeTwoStringSlices(limit int, a, b []string) []string {
+	a = truncateToLimit(limit, a)
+	b = truncateToLimit(limit, b)
+
 	maxl := len(a)
 	if len(b) > len(a) {
 		maxl = len(b)
 	}
+
 	res := make([]string, 0, maxl*10/9)
 
 	for len(a) > 0 && len(b) > 0 {
@@ -56,5 +62,13 @@ func mergeTwoStringSlices(a, b []string) []string {
 	// Append all remaining elements.
 	res = append(res, a...)
 	res = append(res, b...)
+	res = truncateToLimit(limit, res)
 	return res
+}
+
+func truncateToLimit(limit int, a []string) []string {
+	if limit > 0 && len(a) > limit {
+		return a[:limit]
+	}
+	return a
 }