<!--
    Keep PR title verbose enough and add prefix telling
    about what components it touches e.g "query:" or ".*:"
-->

<!--
    Don't forget about CHANGELOG!

    Changelog entry format:
    - [#<PR-id>](<PR-URL>) Thanos <Component> ...

    <PR-id> Id of your pull request.
    <PR-URL> URL of your PR such as https://github.com/thanos-io/thanos/pull/<PR-id>
    <Component> Component affected by your changes such as Query, Store, Receive.
-->

* [ ] I added CHANGELOG entry for this change.
* [X] Change is not relevant to the end user.

## Changes
`MergeSlices` and `MergeUnsortedSlices` now accepts a limit param which will make the merge faster when a lmit is passed.

## Benchmark
```
(base) ➜  thanos git:(merge_slices_with_limit) ✗ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/thanos-io/thanos/pkg/strutil
cpu: Apple M1 Pro
                         │      old.txt      │                  new.txt                  │
                         │      sec/op       │      sec/op        vs base                │
MergeSlices/benchmark-10   0.00275000n ± 13%   0.00006540n ± 16%  -97.62% (p=0.000 n=10)

                         │  old.txt   │            new.txt             │
                         │    B/op    │    B/op     vs base            │
MergeSlices/benchmark-10   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

                         │  old.txt   │            new.txt             │
                         │ allocs/op  │ allocs/op   vs base            │
MergeSlices/benchmark-10   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

<!-- Enumerate changes you made -->

## Verification

<!-- How you tested it? How do you know it works? -->
- Added unit tests
- Added benchmarks